Patient delete error message when trying to delete patient with bed

+++ ApiHospital/Controllers/PatientController.cs
@@ -164,6 +164,9 @@ namespace ApiHospital.Controllers
             if (patient == null)
                 return NotFound();
 
+            if (patient.BedId != null)
+                return BadRequest();
+                
             _context.Patients.Remove(patient);
             await _context.SaveChangesAsync();
